move supporting file logic into a bespoke custom hook
